// ems-frontend/src/models/Auth/EmployeeRegistrationModel.ts
export interface EmployeeRegistrationModel {
    name: string;
    email: string;
    mobileNo: string;
    gender: string;
    age: string;
    salary: string;
    address: string;
    zipcode: string;
    position: string;
    userName: string;
    hiringDate: string;
    departmentId: string;
    countryId: string;
    stateId: string;
    cityId: string;
    reportsTo: string;
    image: File | null;
  }
  